Added IdentityService

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -48,12 +49,15 @@ public class AuctionsController: ControllerBase
         }
         return _mapper.Map<AuctionDto>(auction);
     }
+
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         Auction auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as seller
-        auction.Seller = "test";
+
+        auction.Seller = User.Identity.Name;
+
         _context.Auctions.Add(auction);
 
         // Create a message to send to the message brocker
@@ -79,7 +83,8 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -110,7 +115,7 @@ public class AuctionsController: ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller = username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
